requirements.txt
coverage>=5.1
mock>=4.0.2
pytest>=5.4.0
PyYAML>=5.3
rrdtool>=0.1.15
